+++ src/renderer/components/ShowImageDialog.vue
@@ -10,15 +10,31 @@
         <md-icon>delete</md-icon>
         <md-tooltip>Delete Image</md-tooltip>
       </md-button>
+      <md-button style="float:left;" class="md-primary" @click="showImageInFilesystem( image )">
+        <md-icon>file_download</md-icon>
+        <md-tooltip>View Image</md-tooltip>
+      </md-button>
       <md-button style="float:right;" class="md-primary" @click="showDialog = false">
         <md-icon>close</md-icon>
         <md-tooltip>Close Window</md-tooltip>
       </md-button>
-      <md-button style="float:right;" class="md-primary" @click="moveImageView( image, 'next' )">
+      <md-button style="float:right;" class="md-primary" @click="rotateImage( image, 90 )">
+        <md-icon>rotate_right</md-icon>
+        <md-tooltip>Rotate image to the right 90°</md-tooltip>
+      </md-button>
+      <md-button style="float:right;" class="md-primary" @click="rotateImage( image, 180 )">
+        <md-icon>transform</md-icon>
+        <md-tooltip>Rotate image 180°</md-tooltip>
+      </md-button>
+      <md-button style="float:right;" class="md-primary" @click="rotateImage( image, -90 )">
+        <md-icon>rotate_left</md-icon>
+        <md-tooltip>Rotate image to the left 90°</md-tooltip>
+      </md-button>
+      <md-button style="float:right;" class="md-primary md-raised" @click="moveImageView( image, 'next' )">
         <md-icon>arrow_forward</md-icon>
         <md-tooltip>Next Image</md-tooltip>
       </md-button>
-      <md-button style="float:right;" class="md-primary" @click="moveImageView( image, 'prev' )">
+      <md-button style="float:right;" class="md-primary md-raised" @click="moveImageView( image, 'prev' )">
         <md-icon>arrow_back</md-icon>
         <md-tooltip>Previous Image</md-tooltip>
       </md-button>
@@ -30,6 +46,15 @@
 import { mapActions } from 'vuex';
 
 import { EventBus } from '../store/EventBus';
+import { getThumbnailBlob } from '../../utils';
+
+const { remote,shell } = require('electron');//eslint-disable-line
+
+const logger = require('electron-log');
+const sharp = require('sharp');
+const fs = require('fs');
+
+sharp.cache(false);
 
 export default {
   name: 'show-image-dialog',
@@ -42,12 +67,59 @@ export default {
   }),
   methods: {
     ...mapActions([
-      'ImageGallery/deleteImage'
+      'ImageGallery/deleteImage',
+      'ImageGallery/updateImage'
     ]),
+    showImageInFilesystem(image) {
+      const shown = shell.showItemInFolder(image.src);
+
+      if (!shown) {
+        logger.error('Unable to show item in folder');
+        new Notification('Unable to find image...');//eslint-disable-line
+      }
+    },
+    async rotateImage(image, angle) {
+      image.src = image.src.split('?')[0];//eslint-disable-line
+      const me = this;
+
+      try {
+        await sharp(image.src, { useExifOrientation: false })
+          .rotate(angle)
+          .toFile(`${image.src}.tmp`);
+        fs.renameSync(`${image.src}.tmp`, image.src);
+        const thumbnailBlob = await getThumbnailBlob(image.src);
+        const datetime = new Date();
+        image.src = `${image.src}?${datetime.getTime()}`;
+
+        const reader = new FileReader();
+        reader.onloadend = async function() {//eslint-disable-line
+          image.thumbnail = reader.result;
+          await me.$store.dispatch('ImageGallery/updateImage', image.id, { thumbnail: image.thumbnail }); // TODO: fix this.
+        };
+        reader.readAsDataURL(thumbnailBlob);
+      } catch (e) {
+        logger.error(e);
+      }
+    },
     async deleteImage(image) {
-      await this.$store.dispatch('ImageGallery/deleteImage', image.id);
-      this.showDialog = false;
-      this.image = {};
+      const deleteImage = remote.dialog.showMessageBox(
+        remote.getCurrentWindow(),
+        {
+          title: 'Remove Image?',
+          type: 'question',
+          message: 'Are you sure you want to remove this image from the gallery?',
+          buttons: [
+            'No',
+            'Yes'
+          ]
+        }
+      );
+
+      if (deleteImage) {
+        await this.$store.dispatch('ImageGallery/deleteImage', image.id);
+        this.showDialog = false;
+        this.image = {};
+      }
     },
     viewImage(image) {
       this.image = image;
